chore: add spaghettiapi cmd
We've played around with ReTool and custom API, so far we've tested:

1) Reading and displaying a response from the custom API on the frontend
   in retool
2) Uploading a file and parsing it on the API

In both cases we've found it to work fine. There was some initial
confusion around getting the 14 day trial preview functionality to work,
as we signed up multiple times and there was some smart deduplication on
retool side to prevent multiple signups from receiving the same trial.